Ballphysik überarbeitet.

+++ konsolenspiel/src/main/java/entity/Stein.java
@@ -6,6 +6,7 @@ import java.awt.*;
 
 public class Stein extends Entitaet {
 
+
     public Stein(GamePanel gp, int x, int y) {
 
         this.gp = gp;
@@ -15,11 +16,11 @@ public class Stein extends Entitaet {
 
     public void renderSteinImage(Graphics2D g2) {
         g2.setColor(Color.yellow);
-        g2.fillRect(x, y, gp.tileSize, gp.tileSize / 2);
+        g2.fillRect((int) x, (int) y, gp.tileSize, gp.tileSize / 2);
     }
 
-    public int  getPositionX() {return x;}
+    public int  getPositionX() {return (int) x;}
 
-    public int getPositionY() {return y;}
+    public int getPositionY() {return (int) y;}
 
 }
